Merge updates from Halite-III

+++ game_engine/networking/win32/WinConnection.cpp
@@ -85,7 +85,6 @@ WinConnection::WinConnection(const std::string &command, NetworkingConfig config
         CloseHandle(write[PIPE_TAIL]);
         throw NetworkingError("Could not start process");
     }
-    CloseHandle(piProcInfo.hProcess);
     CloseHandle(piProcInfo.hThread);
     read_pipe = read[PIPE_HEAD];
     write_pipe = write[PIPE_TAIL];
@@ -98,7 +97,16 @@ WinConnection::WinConnection(const std::string &command, NetworkingConfig config
  * Destroy the WinConnection, terminating the subprocess if there is one.
  */
 WinConnection::~WinConnection() noexcept {
-    TerminateProcess(process, 0);
+    if (read_pipe != INVALID_HANDLE_VALUE) {
+        CloseHandle(read_pipe);
+    }
+    if (write_pipe != INVALID_HANDLE_VALUE) {
+        CloseHandle(write_pipe);
+    }
+    if (process != INVALID_HANDLE_VALUE) {
+        TerminateProcess(process, 0);
+        CloseHandle(process);
+    }
 }
 
 /**
@@ -126,7 +134,7 @@ std::string WinConnection::get_string(std::chrono::milliseconds timeout) {
     auto initial_time = high_resolution_clock::now();
     while (true) {
         DWORD bytes_available = 0;
-        while (bytes_available < 1) {
+        while (true) {
             if (!config.ignore_timeout) {
                 auto current_time = high_resolution_clock::now();
                 auto remaining = timeout - duration_cast<milliseconds>(current_time - initial_time);
@@ -134,7 +142,14 @@ std::string WinConnection::get_string(std::chrono::milliseconds timeout) {
                     throw TimeoutError("when reading string", timeout, result);
                 }
             }
+
             PeekNamedPipe(read_pipe, nullptr, 0, nullptr, &bytes_available, nullptr);
+
+            if (bytes_available > 0) {
+                break;
+            }
+
+            Sleep(1);
         }
 
         DWORD chars_read;
